Implement download clients, fix bugs
