+++ buildarr_prowlarr/config/settings/download_clients/__init__.py
@@ -13,24 +13,49 @@
 
 
 """
-Sonarr plugin download client settings.
+Prowlarr plugin download client settings.
 """
 
 
 from __future__ import annotations
 
 from logging import getLogger
-from typing import Dict, List, Mapping, Union
+from typing import Dict, Tuple, Type, Union
 
-from buildarr.config import RemoteMapEntry
+import prowlarr
+
 from pydantic import Field
 from typing_extensions import Annotated, Self
 
-from ...api import api_get, api_put
-from ...secrets import SonarrSecrets
-from ...types import SonarrConfigBase
-from .download_clients import (
-    DOWNLOADCLIENT_TYPE_MAP,
+from ....api import prowlarr_api_client
+from ....secrets import ProwlarrSecrets
+from ...types import ProwlarrConfigBase
+from .base import DownloadClient
+from .torrent import (
+    Aria2DownloadClient,
+    DelugeDownloadClient,
+    DownloadstationTorrentDownloadClient,
+    FloodDownloadClient,
+    HadoukenDownloadClient,
+    QbittorrentDownloadClient,
+    RtorrentDownloadClient,
+    TorrentBlackholeDownloadClient,
+    TransmissionDownloadClient,
+    UtorrentDownloadClient,
+    VuzeDownloadClient,
+)
+from .usenet import (
+    DownloadstationUsenetDownloadClient,
+    NzbgetDownloadClient,
+    NzbvortexDownloadClient,
+    PneumaticDownloadClient,
+    SabnzbdDownloadClient,
+    UsenetBlackholeDownloadClient,
+)
+
+logger = getLogger(__name__)
+
+DOWNLOADCLIENT_TYPES: Tuple[Type[DownloadClient], ...] = (
     Aria2DownloadClient,
     DelugeDownloadClient,
     DownloadstationTorrentDownloadClient,
@@ -49,9 +74,11 @@ from .download_clients import (
     UtorrentDownloadClient,
     VuzeDownloadClient,
 )
-from .remote_path_mappings import SonarrRemotePathMappingsSettingsConfig
 
-logger = getLogger(__name__)
+DOWNLOADCLIENT_TYPE_MAP = {
+    downloadclient_type._implementation.lower(): downloadclient_type
+    for downloadclient_type in DOWNLOADCLIENT_TYPES
+}
 
 DownloadClientType = Union[
     DownloadstationUsenetDownloadClient,
@@ -74,47 +101,27 @@ DownloadClientType = Union[
 ]
 
 
-class SonarrDownloadClientsSettingsConfig(SonarrConfigBase):
+class ProwlarrDownloadClientsSettings(ProwlarrConfigBase):
     """
-    Download clients retrieve media files being tracked by Sonarr,
-    and store them in a location Sonarr can access to manage the
+    Download clients retrieve media files being tracked by Prowlarr,
+    and store them in a location Prowlarr can access to manage the
     downloaded files.
 
-    Download clients that use Usenet or BitTorrent can be configured,
-    as well as remote path mappings and other related options.
+    Download clients that use Usenet or BitTorrent can be configured.
 
     ```yaml
     ---
 
-    sonarr:
+    prowlarr:
       settings:
         download_clients:
-          enable_completed_download_handling: true
-          redownload_failed: true
-          delete_unmanaged: true
           definitions:
             Transmission:
               type: "transmission"
               host: "transmission"
               port: 9091
             ...
-          remote_path_mappings:
-            definitions:
-              - host: "transmission"
-                remote_path: "/remote/path"
-                local_path: "/local/path"
-              ...
     ```
-    """
-
-    enable_completed_download_handling: bool = True
-    """
-    Automatically import completed downloads from download clients.
-    """
-
-    redownload_failed: bool = True
-    """
-    Automatically search for and attempt to download a different release.
     """
 
     delete_unmanaged: bool = False
@@ -127,138 +134,85 @@ class SonarrDownloadClientsSettingsConfig(SonarrConfigBase):
     Download client definitions, for connecting with external media downloaders.
     """
 
-    remote_path_mappings = SonarrRemotePathMappingsSettingsConfig()
-    """
-    Configuration for mapping paths on download client hosts to their counterparts
-    on this Sonarr instance.
-
-    For more information, refer to "Configuring remote path mappings".
-    """
-
-    _remote_map: List[RemoteMapEntry] = [
-        ("enable_completed_download_handling", "enableCompletedDownloadHandling", {}),
-        ("redownload_failed", "autoRedownloadFailed", {}),
-    ]
-
     @classmethod
-    def from_remote(cls, secrets: SonarrSecrets) -> Self:
-        downloadclient_config = api_get(secrets, "/api/v3/config/downloadclient")
-        downloadclients = api_get(secrets, "/api/v3/downloadclient")
-        tag_ids: Dict[str, int] = (
-            {tag["label"]: tag["id"] for tag in api_get(secrets, "/api/v3/tag")}
-            if any(downloadclient["tags"] for downloadclient in downloadclients)
-            else {}
-        )
+    def from_remote(cls, secrets: ProwlarrSecrets) -> Self:
+        with prowlarr_api_client(secrets=secrets) as api_client:
+            api_downloadclients = prowlarr.DownloadClientApi(api_client).list_download_client()
+            tag_ids: Dict[str, int] = (
+                {tag.label: tag.id for tag in prowlarr.TagApi(api_client).list_tag()}
+                if any(api_downloadclient.tags for api_downloadclient in api_downloadclients)
+                else {}
+            )
         return cls(
-            **cls.get_local_attrs(cls._remote_map, downloadclient_config),
             definitions={
-                dc["name"]: DOWNLOADCLIENT_TYPE_MAP[dc["implementation"]]._from_remote(
+                api_downloadclient.name: DOWNLOADCLIENT_TYPE_MAP[
+                    api_downloadclient.implementation.lower()
+                ]._from_remote(
                     tag_ids=tag_ids,
-                    remote_attrs=dc,
+                    remote_attrs=api_downloadclient.to_dict(),
                 )
-                for dc in downloadclients
+                for api_downloadclient in api_downloadclients
             },
-            remote_path_mappings=SonarrRemotePathMappingsSettingsConfig._from_remote(
-                secrets=secrets,
-            ),
         )
 
     def update_remote(
         self,
         tree: str,
-        secrets: SonarrSecrets,
+        secrets: ProwlarrSecrets,
         remote: Self,
         check_unmanaged: bool = False,
     ) -> bool:
-        # Update download client-related configuration options.
-        config_updated, config_remote_attrs = self.get_update_remote_attrs(
-            tree,
-            remote,
-            self._remote_map,
-            check_unmanaged=check_unmanaged,
-        )
-        if config_updated:
-            remote_config = api_get(secrets, "/api/v3/config/downloadclient")
-            api_put(
-                secrets,
-                f"/api/v3/config/downloadclient/{remote_config['id']}",
-                {
-                    "id": remote_config["id"],
-                    "downloadClientWorkingFolders": remote_config["downloadClientWorkingFolders"],
-                    **config_remote_attrs,
-                },
-            )
-        # Update download clients.
-        definitions_updated = self._update_remote_definitions(
-            tree=f"{tree}.definitions",
-            secrets=secrets,
-            local=self.definitions,
-            remote=remote.definitions,
-            check_unmanaged=check_unmanaged,
-        )
-        # Update remote path mappings.
-        rpms_updated = self.remote_path_mappings._update_remote(
-            tree=f"{tree}.remote_path_mappings",
-            secrets=secrets,
-            remote=remote.remote_path_mappings,
-        )
         #
-        return any([config_updated, definitions_updated, rpms_updated])
-
-    def _update_remote_definitions(
-        self,
-        tree: str,
-        secrets: SonarrSecrets,
-        local: Mapping[str, DownloadClientType],
-        remote: Mapping[str, DownloadClientType],
-        check_unmanaged: bool,
-    ) -> bool:
         changed = False
-        downloadclient_ids: Dict[str, int] = {
-            downloadclient_json["name"]: downloadclient_json["id"]
-            for downloadclient_json in api_get(secrets, "/api/v3/downloadclient")
-        }
-        tag_ids: Dict[str, int] = (
-            {tag["label"]: tag["id"] for tag in api_get(secrets, "/api/v3/tag")}
-            if any(downloadclient.tags for downloadclient in local.values())
-            or any(downloadclient.tags for downloadclient in remote.values())
-            else {}
-        )
-        # Create download clients that don't exist yet on the remote,
-        # and update in-place ones that do.
-        for downloadclient_name, downloadclient in local.items():
-            downloadclient_tree = f"{tree}[{repr(downloadclient_name)}]"
-            if downloadclient_name not in remote:
+        #
+        with prowlarr_api_client(secrets=secrets) as api_client:
+            downloadclient_api = prowlarr.DownloadClientApi(api_client)
+            api_downloadclient_schemas = downloadclient_api.list_download_client_schema()
+            api_downloadclients = {
+                api_downloadclient.name: api_downloadclient
+                for api_downloadclient in downloadclient_api.list_download_client()
+            }
+            tag_ids: Dict[str, int] = (
+                {tag.label: tag.id for tag in prowlarr.TagApi(api_client).list_tag()}
+                if any(downloadclient.tags for downloadclient in self.definitions.values())
+                or any(downloadclient.tags for downloadclient in remote.definitions.values())
+                else {}
+            )
+        #
+        for downloadclient_name, downloadclient in self.definitions.items():
+            downloadclient_tree = f"{tree}.definitions[{repr(downloadclient_name)}]"
+            #
+            if downloadclient_name not in remote.definitions:
                 downloadclient._create_remote(
                     tree=downloadclient_tree,
                     secrets=secrets,
+                    api_downloadclient_schemas=api_downloadclient_schemas,
                     tag_ids=tag_ids,
                     downloadclient_name=downloadclient_name,
                 )
                 changed = True
+            #
             elif downloadclient._update_remote(
                 tree=downloadclient_tree,
                 secrets=secrets,
-                remote=remote[downloadclient_name],  # type: ignore[arg-type]
+                remote=remote.definitions[downloadclient_name],  # type: ignore[arg-type]
+                api_downloadclient_schemas=api_downloadclient_schemas,
                 tag_ids=tag_ids,
-                downloadclient_id=downloadclient_ids[downloadclient_name],
-                downloadclient_name=downloadclient_name,
+                api_downloadclient=api_downloadclients[downloadclient_name],
             ):
                 changed = True
-        # If `delete_unmanaged` is `True`, remove any download clients on the remote
-        # that aren't managed by Buildarr.
-        # Otherwise, just log them.
-        for downloadclient_name, downloadclient in remote.items():
-            if downloadclient_name not in local:
-                downloadclient_tree = f"{tree}[{repr(downloadclient_name)}]"
+        #
+        for downloadclient_name, downloadclient in remote.definitions.items():
+            if downloadclient_name not in self.definitions:
+                downloadclient_tree = f"{tree}.definitions[{repr(downloadclient_name)}]"
                 if self.delete_unmanaged:
                     downloadclient._delete_remote(
                         tree=downloadclient_tree,
                         secrets=secrets,
-                        downloadclient_id=downloadclient_ids[downloadclient_name],
+                        downloadclient_id=api_downloadclients[downloadclient_name].id,
                     )
                     changed = True
                 else:
                     logger.debug("%s: (...) (unmanaged)", downloadclient_tree)
-        # Return the resource changed status.
+        #
         return changed
